refactor(app): move ui from shared to application/shared
