+++ libs/application/shared/ui/focus/README.md
@@ -1,0 +1,7 @@
+# application/shared/ui/focus
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test application/shared/ui/focus` to execute the unit tests.
