+++ libs/application/shared/ui/modal-header/README.md
@@ -1,0 +1,7 @@
+# application/shared/ui/modal-header
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test application/shared/ui/modal-header` to execute the unit tests.
