+++ libs/application/shared/ui/modal-header/src/lib/modal-header.directive.ts
