+++ libs/application/shared/ui/page-header/README.md
@@ -1,0 +1,7 @@
+# application/shared/ui/page-header
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test application/shared/ui/page-header` to execute the unit tests.
