+++ libs/application/shared/ui/page-header/tsconfig.lib.json
@@ -1,7 +1,7 @@
 {
   "extends": "./tsconfig.json",
   "compilerOptions": {
-    "outDir": "../../../../dist/out-tsc",
+    "outDir": "../../../../../dist/out-tsc",
     "target": "es2015",
     "declaration": true,
     "declarationMap": true,
