--- libs/shared/ui/focus/README.md
@@ -1,7 +1,0 @@
-# shared-ui-focus
-
-This library was generated with [Nx](https://nx.dev).
-
-## Running unit tests
-
-Run `nx test shared-ui-focus` to execute the unit tests.
