--- libs/shared/ui/modal-header/README.md
@@ -1,7 +1,0 @@
-# shared-ui-modal-header
-
-This library was generated with [Nx](https://nx.dev).
-
-## Running unit tests
-
-Run `nx test shared-ui-modal-header` to execute the unit tests.
