--- libs/shared/ui/page-header/README.md
@@ -1,7 +1,0 @@
-# shared-ui-page-header
-
-This library was generated with [Nx](https://nx.dev).
-
-## Running unit tests
-
-Run `nx test shared-ui-page-header` to execute the unit tests.
